extensions refactoring; reorganized, old files removed temporary references added to other files to maintain functionality before namespace trim
