+++ ULTRAKIT.Extensions/AssetLoader.cs
@@ -1,73 +1,77 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using UnityEngine;
-using UnityEditor;
 
-namespace ULTRAKIT.Extensions
+namespace ULTRAKIT.Extensions.Extensions
 {
     public static class AssetLoader
     {
-        public static object LoadAsset(string filePath, string name)
-        {
-            AssetBundle bundle = AssetBundle.LoadFromFile(filePath);
-            object asset = bundle.LoadAsset(name);
-
-            Debug.Log($"Loaded asset {name} from {filePath}");
-            return asset;
-        }
-
-        public static object[] LoadAllAssets(string filePath)
-        {
-            AssetBundle bundle = AssetBundle.LoadFromFile(filePath);
-            object[] assets = bundle.LoadAllAssets();
-            return assets;
-        }
-
-        public static object[] LoadAssets(string filePath, string[] names)
-        {
-            AssetBundle bundle = AssetBundle.LoadFromFile(filePath);
-            List<object> assets = new List<object>();
-
-            foreach (string name in names)
-            {
-                assets.Add(bundle.LoadAsset(name));
-            }
-
-            return assets.ToArray();
-        }
-
-        public static AssetBundleRequest LoadAssetAsync(string filePath, string name)
-        {
-            AssetBundle bundle = AssetBundle.LoadFromFile(filePath);
-            return bundle.LoadAssetAsync(name);
-        }
-
-        public static AssetBundleRequest LoadAllAssetsAsync(string filePath)
-        {
-            AssetBundle bundle = AssetBundle.LoadFromFile(filePath);
-            return bundle.LoadAllAssetsAsync();
-        }
-
-        public static List<AssetBundleRequest> LoadAssetsAsync(string filePath, string[] names)
-        {
-            AssetBundle bundle = AssetBundle.LoadFromFile(filePath);
-            List<AssetBundleRequest> requests = new List<AssetBundleRequest>();
-
-            foreach (string name in names)
-            {
-                requests.Add(bundle.LoadAssetAsync(name));
-            }
-
-            return requests;
-        }
-
         public static T[] GetAll<T>(object[] allAssets)
         {
             T[] assets = (T[])allAssets.Where(k => k is T).ToArray().Cast<T>();
             return assets;
+        }
+
+        public static T AssetFind<T>(string name) where T : UnityEngine.Object
+        {
+            T result = default;
+            T[] allAssets = Resources.FindObjectsOfTypeAll<T>();
+
+            foreach (T asset in allAssets)
+                if (asset.name == name) return result;
+            UKLogger.LogWarning($"Could not find asset {name}");
+            return result;
+        }
+
+        public static T AssetFind<T>(this AssetBundle bundle, string name) where T : UnityEngine.Object
+        {
+            return bundle.LoadAsset<T>(name) ?? default;
+        }
+
+        public static T AssetFind<T>(string bundleName, string name, bool searchActBundles = false) where T : UnityEngine.Object
+        {
+            AssetBundle bundle;
+            if (LoadFromLoaded(bundleName, out bundle))
+            {
+                return bundle.LoadAsset<T>(name) ?? default;
+            }
+
+            string target = $@"{Application.productName}_Data\StreamingAssets\{bundleName}";
+            string target2 = $@"{Application.productName}_Data\StreamingAssets\acts\{bundleName}";
+            if (File.Exists(target))
+            {
+                var data = File.ReadAllBytes(target);
+                bundle = AssetBundle.LoadFromMemory(data);
+                return bundle.LoadAsset<T>(name) ?? default;
+            }
+            if (File.Exists(target2) && searchActBundles)
+            {
+                var data = File.ReadAllBytes(target2);
+                bundle = AssetBundle.LoadFromMemory(data);
+                return bundle.LoadAsset<T>(name) ?? default;
+            }
+
+            UKLogger.LogWarning($"Could not find bundle {bundleName} or StreamingAssets file");
+            return default;
+        }
+
+        public static bool LoadFromLoaded(string name, out AssetBundle result)
+        {
+            result = null;
+            IEnumerable<AssetBundle> loadedBundles = AssetBundle.GetAllLoadedAssetBundles();
+
+            foreach (AssetBundle bundle in loadedBundles)
+                if (bundle.name == name)
+                {
+                    result = bundle;
+                    return true;
+                }
+
+            return false;
         }
     }
 }
