+++ ULTRAKIT.Loader/Loaders/SpawnerArmLoader.cs
@@ -6,6 +6,7 @@ using System.Text;
 using System.Threading.Tasks;
 using ULTRAKIT.Data;
 using ULTRAKIT.Extensions;
+using ULTRAKIT.Extensions.ObjectClasses.Monobehaviours;
 using UnityEngine;
 
 namespace ULTRAKIT.Loader
